:truck: Move adminsitelog to contrib.admin_tools and guard missing admin
Relocate the adminsitelog command into the opt-in
django_boost.contrib.admin_tools app, matching its hard dependency on
django.contrib.admin. The former django_boost command is kept as a
re-export shim that raises a DeprecationWarning and will be removed in
django-boost 4.0. adminsitelog now defers its LogEntry import and exits
with a clear CommandError when django.contrib.admin is not installed,
instead of an unrelated model-loading RuntimeError.
